+++ lib/unicode/property/word_break.ex
@@ -12,8 +12,10 @@ defmodule Unicode.WordBreak do
 
   alias Unicode.Utils
 
+  # `@missing: 0000..10FFFF; Other` - see `Utils.add_default_value/2`.
   @word_breaks Utils.word_breaks()
                |> Utils.remove_annotations()
+               |> Utils.add_default_value(:other)
 
   @word_break_table Unicode.RangeSearch.new_value_table(@word_breaks)